backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomqtt>=2.4.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.122.0",
    "fastapi-mail>=1.5.8",
    "pwdlib[argon2]>=0.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
